fix: channel not getting created and added flag for channel creation and refactoring
