+++ sg_slack_integration/doc_events/project.py
@@ -1,5 +1,6 @@
 from sg_slack_integration.doc_events.common_function import create_slack_channel,set_topic,set_description,get_channel_id
 
+
 def validate(self, method=None):
     create_slack_channel(self)
     set_topic_and_description(self)
